--- src/components/Hub/Hub.js
@@ -1,12 +1,0 @@
-import BackgroundImage from 'components/BackgroundImage/BackgroundImage';
-import NavBar from 'components/NavBar/NavBar';
-import MainContainer from 'components/MainContainer/MainContainer';
-
-const Hub = () => (
-  <BackgroundImage>
-    <NavBar />
-    <MainContainer />
-  </BackgroundImage>
-);
-
-export default Hub;
