+++ src/components/background/index.js
