+++ src/components/characterMarquee/marqueeStats/main/index.js
