+++ src/components/characterMarquee/marqueeStats/regen/index.js
