+++ src/components/equipmentList/index.js
