+++ src/components/hubTabs/tabDetail/index.js
